Relocate the TagBox Component
