+++ src/components/AssignmentBox.vue
@@ -31,7 +31,7 @@
 import { Component, Prop, Vue } from "vue-property-decorator";
 import { Assignment } from "@/types/garnbarn/Assignment";
 import DialogBox from "@/components/DialogBox/DialogBox";
-import TagBox from "@/components/TagBox.vue";
+import TagBox from "@/components/Tag/TagBox.vue";
 
 @Component({
   components: {
