+++ tests/unit/TagBox.spec.ts
@@ -1,5 +1,5 @@
 import { createLocalVue, mount, Wrapper } from "@vue/test-utils";
-import TagBox from "@/components/TagBox.vue";
+import TagBox from "@/components/Tag/TagBox.vue";
 import { Tag } from "@/types/garnbarn/Tag";
 import VueMaterial from "vue-material";
